Rename module from notify to push
